src/scraper/requirements.txt
azure-identity>=1.19.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.28.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
uvicorn>=0.34.0
fastapi>=0.115.0
azure-servicebus>=7.13.0
azure-storage-blob>=12.24.0
azure-cosmos>=4.9.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
